BOL-227 refactor : 겹치는 파일 develop 에 있는 것으로 수정

+++ presentation/src/main/java/com/yapp/bol/presentation/viewmodel/login/MyGroupList.kt
@@ -13,7 +13,7 @@ object MyGroupList {
         return myGroupList
     }
 
-    fun findMyGroup(groupId: Int?): MyGroupItem? {
-        return myGroupList?.find { it.id == groupId }
+    fun findMyGroup(groupId: Long?): MyGroupItem? {
+        return myGroupList?.find { it.id.toLong() == groupId }
     }
 }
